added the essential functions + add_guest loop
I added the personalized and customizable functions, the invitation function, the set dinner time function, as well as making it so that there is a loop to add the max guests, it keeps appending the guests until it reaches the max amount. I used len in my code to print out the guests neatly with numbers.